+++ flatlaf-core/src/test/java/com/formdev/flatlaf/ui/TestFlatStyleableInfo.java
@@ -398,7 +398,8 @@ public class TestFlatStyleableInfo
 
 		Map<String, Class<?>> expected = new LinkedHashMap<>();
 		expectedMap( expected,
-			"showCapsLock", boolean.class
+			"showCapsLock", boolean.class,
+			"showRevealButton", boolean.class
 		);
 
 		// FlatPasswordFieldUI extends FlatTextFieldUI