+++ flatlaf-testing/src/main/java/com/formdev/flatlaf/testing/FlatTextComponentsTest.java
@@ -126,6 +126,14 @@ public class FlatTextComponentsTest
 			showClearButtonCheckBox.isSelected() );
 	}
 
+	private void showRevealButton() {
+		for( Component c : getComponents() ) {
+			if( c instanceof JPasswordField )
+				((JPasswordField)c).putClientProperty(FlatClientProperties.STYLE,
+					showRevealButtonCheckBox.isSelected() ? "showRevealButton: true" : null );
+		}
+	}
+
 	private void putTextFieldClientProperty( String key, Object value ) {
 		for( Component c : getComponents() ) {
 			if( c instanceof JTextField )
@@ -174,6 +182,7 @@ public class FlatTextComponentsTest
 		leadingComponentVisibleCheckBox = new JCheckBox();
 		trailingComponentVisibleCheckBox = new JCheckBox();
 		showClearButtonCheckBox = new JCheckBox();
+		showRevealButtonCheckBox = new JCheckBox();
 		JLabel passwordFieldLabel = new JLabel();
 		JPasswordField passwordField1 = new JPasswordField();
 		JPasswordField passwordField3 = new JPasswordField();
@@ -326,6 +335,7 @@ public class FlatTextComponentsTest
 				"[]" +
 				"[]0" +
 				"[]" +
+				"[]" +
 				"[]"));
 
 			//---- button1 ----
@@ -417,6 +427,12 @@ public class FlatTextComponentsTest
 			showClearButtonCheckBox.setName("showClearButtonCheckBox");
 			showClearButtonCheckBox.addActionListener(e -> showClearButton());
 			panel1.add(showClearButtonCheckBox, "cell 0 11 2 1,alignx left,growx 0");
+
+			//---- showRevealButtonCheckBox ----
+			showRevealButtonCheckBox.setText("password reveal button");
+			showRevealButtonCheckBox.setName("showRevealButtonCheckBox");
+			showRevealButtonCheckBox.addActionListener(e -> showRevealButton());
+			panel1.add(showRevealButtonCheckBox, "cell 0 12 2 1,alignx left,growx 0");
 		}
 		add(panel1, "cell 4 0 1 10,aligny top,growy 0");
 
@@ -733,6 +749,7 @@ public class FlatTextComponentsTest
 	private JCheckBox leadingComponentVisibleCheckBox;
 	private JCheckBox trailingComponentVisibleCheckBox;
 	private JCheckBox showClearButtonCheckBox;
+	private JCheckBox showRevealButtonCheckBox;
 	private JTextField textField;
 	private JCheckBox dragEnabledCheckBox;
 	private JTextArea textArea;